pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "order-subscription-service"
version = "0.1.0"
description = ""
authors = [
  {name = "bodrstvuyushiy", email = "grapegregory@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
  "django (==5.2.1)",
  "djangorestframework (==3.14.0)",
  "psycopg2-binary (>=2.9.9,<3.0.0)",
  "requests (>=2.32.5,<3.0.0)",
  "celery (>=5.3,<5.4)",
  "redis (>=5.0,<6.0)",
  "django-redis (>=5.4,<5.6)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-aaa = "^0.17.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-mock = "^3.14.0"
ruff = "^0.7.0"
toml-sort = "^0.23.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py"]
testpaths = ["tests"]
pythonpath = ["src"]
